src/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.2.0
python-dotenv==0.20.0
pytz==2022.2
six==1.16.0
Werkzeug==2.2.2
